単語帳部分のモード切替
