+++ app/create/page.tsx
@@ -2,7 +2,6 @@
 //import
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
-import { Checkbox } from "@/components/ui/checkbox";
 import Card from "@/components/ui-elements/card";
 import { useState } from "react";
 import { v4 as createUUID } from "uuid";
